# Description

This PR improves the http file cache performance by avoiding retriving the metadata information if the file cache is present.

Solves #4010 

# Performance number:
We run the same query twice in the same transaction:
`load from 's3://kuzu-test/dataset/ldbc10/comment_0_0.csv'(header=true,delim='|') return *;`
LDBC10: 2.1 GB
This branch:
| Run ID | compiling time | execution time | total time |
| -------- | ------- |------- |------- |
| 1 | 27.32s | 0.846s | 28.166 s |
| 2 | < 1ms | 0.812s | 0.812 s |


Master:
| Run ID | compiling time | execution time | total time |
| -------- | ------- |------- |------- |
| 1 | 27.05s | 4.59s | 31.64 s |
| 2 | 0.249s | 4.367s | 5.616 s |